Evokes the Application class of the gem's module
This app uses version 0.1.4 of myfirstgem. In this version the call to
Application.new in the config.ru file passes an instance of this
application (using_my_gem) and upon request from the browser the gem
responds with a text/html message.

It's also good to note this web application can respond to a single
request without the use of ActionView, ActiveRecord, ActiveModel,
controller, models or views.

This is a basic use of the puma web server and Rack.